Enhance synoptic data from Purple Air

Enhance raw synoptic data from PurpleAir to create a generally useful dataframe.

Steps include:

1) Replace variable with more consistent, more human readable names.

2) Add spatial metadata for each monitor including:

timezone -- olson timezone

countryCode -- ISO 3166-1 alpha-2

stateCode -- ISO 3166-2 alpha-2

3) Convert data types from character to POSIXct and numeric.

4) Add distance and monitorID for the two closest PWFSL monitors

Subsetting by country may be performed by specifying the countryCodes argument.

Setting outsideOnly = TRUE will return only those records marked as 'outside'.

enhanceSynopticData(pas_raw, countryCodes = c("US"),
  includePWFSL = TRUE)

Arguments

pas_raw
Dataframe returned by downloadParseSynopticData().

countryCodes
ISO country codes used to subset the data.

includePWFSL
Logical specifying whether to calculate distances from PWFSL monitors.

Value

Enhanced Dataframe of synoptic PurpleAir data.

Note

For data obtained on July 28, 2018 this will result in removal of all 'B' channels, even those whose parent 'A' channel is marked as 'outside'. This is useful if you want a quick, synoptic view of the network, e.g. for a map.

See also

Examples

# NOT RUN {
initializeMazamaSpatialUtils()
pas_raw <- downloadParseSynopticData()
pas <- enhanceSynopticData(pas_raw)
# }